+++ lib/r2.rb
@@ -2,7 +2,9 @@
 
 require "r2/cli"
 require "r2/configuration"
+require "r2/content_type"
 require "r2/errors"
 require "r2/logging"
+require "r2/retry"
 require "r2/storage"
 require "r2/version"